+++ src/DataBar/demo/columns.tsx
@@ -45,7 +45,7 @@ const getStyleByValue = ({
   if (typeof value !== 'number' || isNaN(value)) {
     return {};
   }
-  if (value < min || value > max) {
+  if (min > max) {
     return {};
   }
   const absValue = Math.abs(value);
@@ -60,6 +60,9 @@ const getStyleByValue = ({
   const negativePureColor = negativeGradient[0];
   // 只有正数情况
   if (min > 0) {
+    if (value < min) {
+      return {};
+    }
     const { background, border } = getLinearGradientStyle({
       colors: positiveGradient,
       direction: 'to right',
@@ -73,6 +76,9 @@ const getStyleByValue = ({
   }
   // 只有负数情况
   if (max < 0) {
+    if (value > max) {
+      return {};
+    }
     const { background, border } = getLinearGradientStyle({
       colors: negativeGradient,
       direction: 'to left',
@@ -93,7 +99,7 @@ const getStyleByValue = ({
       direction: 'to right',
     });
     // 正值宽度百分比 = 值在正值范围的占比 * 正值范围在总范围的占比 * 100%
-    const valueWidth = (value / max) * (max / (absMin + absMax)) * 100;
+    const valueWidth = (Math.min(value, max) / max) * (max / (absMin + absMax)) * 100;
     return {
       left: zeroPosition + '%',
       width: valueWidth + '%',
@@ -107,7 +113,7 @@ const getStyleByValue = ({
       direction: 'to left',
     });
     // 负值宽度百分比 = 值在负值范围的占比 * 负值范围在总范围的占比 * 100%
-    const valueWidth = (absValue / absMin) * (absMin / (absMin + absMax)) * 100;
+    const valueWidth = (Math.min(absValue, absMin) / absMin) * (absMin / (absMin + absMax)) * 100;
     return {
       right: 100 - zeroPosition + '%',
       width: valueWidth + '%',
@@ -202,15 +208,15 @@ export function getColumns({
     },
 
     {
-      title: t('demo.table.fixedNegativePure'),
+      title: t('demo.table.fixedNegativePure', { min, max }),
       dataIndex: 'negativeValue',
       render(value) {
-        // const curVal = calculateValues(dataSource, 'negativeValue');
+        const curVal = calculateValues(dataSource, 'negativeValue');
         const style = getStyleByValue({
           ...styleByValueParams,
           value,
-          max: -1,
-          min: -10,
+          max: max ?? curVal.max,
+          min: min ?? curVal.min,
           isGradient: false,
         });
         return (